chore: use linter and move functions
